I christen this game a full fledged port
The static image of Zack in previous attempts at a standalone port served mostly as a placeholder because I haven't animanted a sprite with a general purpose game engine and it's designer since Game Maker 6. This was no different.

But not anymore! The eyes are animanted now and soon the mouth will be too. I know that sounds small in the grand scheme of things but that pretty much means that the design of this is identical to the original Flash version. So I can say with confidence that this pretty much counts as a full fledged port of the original. I'm so glad I never gave up on figuring out how to bring this game to a wider audience in it's entirety. Think I'll be sticking with Godot for future games now.