Zack is now Nathan
The character in this dress up game is now a cat and named Nathan. This is because my original style unintentionally made him look like a cat. His name change to Nathan from Zack is because I don't want to cause confusion between people who use to know me as Zack and my first fursona.