SVG remaster, part 1 (#5)
Up until this point, this game still relied heavily on my copy of ZC's Dress Up for editing and creating new sprites so they properly match the look and feel. The only reason why it's not included in the source code is because the format is proprietary and Adobe's editor doesn't work with all platforms that Godot does. That is unfair when the whole goal of this project to remove Flash from the equation.

This remaster aims to remove that final need by remaking in an open format, whether it be an open source editor or format. I found that using SVG as is with Inkscape as the editor does the job well because Godot can render it without any extra steps required. I tried testing out Krita and Gimp but they didn't produce the results I wanted.