Switch to C# (#8)
I learned C# primarily for game development but held off switching to it on Godot 3 because it was still rough around edges - crashed way too many times. It seems to be working as it should as of 3.1.1 and that's why I'm rewriting all the GDScript portions in C# as much as possible.

However, this does make it incompatible with the Steam version, as of this writing, until the Godot devs decide how they want to publish it in future updates.