DLC support (#11)
This update adds support for loading in DLC using the check_for_dlc() function. New clothes and accessories are added by searching through "res://scenes/dlc/shirts/" directory, for example, and added to the respective tab in the wardrobe as if it were part of the base game. It's not the best solution in the world but it works.