Modular character design (#12)
Legs, shirt and the head are now independent sprites that can be changed separately. New clothes can be made from a template instead of having to work within a blank space. This removes the need for extra layers and placeholders while opening the door to creating new scenes and migrating over to new styles.

This wasn't always possible until the SVG port removed the Adobe Animate dependency. The layers and placeholders were there because the original Flash game was based around a drag-and-drop design. Tony's Dress Up was initially intended to be a 1:1 port of the Flash game which is why it wasn't done sooner.

New Emoji-based shirts based on Twemoji because the idea sounded fun. Credits will eventually be added in-game but for now it's in the README. For the time being, a few of the ZC's Dress Up-era shirts have pulled because they were never properly ported over during the remaster.